+++ src/bookmark_reader/mod.rs
@@ -13,30 +13,49 @@ pub use simple::SimpleBookmarkReader;
 pub use source_reader::SourceReader;
 use std::{
     fmt,
-    fs::File,
     io::Read,
     path::{Path, PathBuf},
 };
 pub use target_reader::ReadTarget;
 pub use target_writer::WriteTarget;
 
+#[derive(Debug, PartialEq)]
+pub enum ReaderName {
+    Firefox,
+    FirefoxCompressed,
+    Chromium,
+    ChromiumNoExtension,
+    Simple,
+}
+
+impl fmt::Display for ReaderName {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let reader_name = match &self {
+            ReaderName::Firefox => "Firefox",
+            ReaderName::FirefoxCompressed => "Firefox (compressed)",
+            ReaderName::Chromium => "Chromium",
+            ReaderName::ChromiumNoExtension => "Chromium (no extension)",
+            ReaderName::Simple => "Simple",
+        };
+        write!(f, "{}", reader_name)
+    }
+}
+
 /// A trait to read bookmarks from multiple sources, like Firefox or Chrome.
 pub trait ReadBookmark: fmt::Debug {
-    fn name(&self) -> &str;
-
-    fn source_type(&self, source_path: &Path) -> Result<SourceType, anyhow::Error>;
+    fn name(&self) -> ReaderName;
 
     fn extension(&self) -> Option<&str>;
 
-    fn is_selected(&self, bookmarks: &str) -> Result<bool, anyhow::Error>;
+    /// Identify and select the source.
+    ///
+    /// A bookmark reader can read from multiple source. For example, the
+    /// standard from Chromium is used for Chrome and Edge as well.
+    fn select_source(&self, source_path: &Path) -> Result<Option<SourceType>, anyhow::Error>;
 
-    fn select_path(&self, _path: &Path) -> Result<Option<PathBuf>, anyhow::Error> {
+    /// Select the bookmarks file if the source is given as a directory.
+    fn select_file(&self, _source_dir: &Path) -> Result<Option<PathBuf>, anyhow::Error> {
         Ok(None)
-    }
-
-    fn open(&self, path: &Path) -> Result<File, anyhow::Error> {
-        let file = File::open(path)?;
-        Ok(file)
     }
 
     fn read(&self, reader: &mut dyn Read) -> Result<String, anyhow::Error> {
@@ -60,7 +79,7 @@ pub trait ReadBookmark: fmt::Debug {
         source: &Source,
         bookmarks: &mut SourceBookmarks,
     ) -> Result<(), anyhow::Error> {
-        debug!("Read bookmarks from file '{}'", source.path.display());
+        debug!("Read bookmarks from file '{}'", source.path);
 
         let raw_bookmarks = self.read(reader)?;
         self.parse(&raw_bookmarks, source, bookmarks)?;