+++ src/cmd/init.rs
@@ -65,7 +65,7 @@ async fn init_bookmarks(
         source_reader.len(),
         source_reader
             .iter()
-            .map(|reader| reader.source().path.to_string_lossy())
+            .map(|reader| reader.source().path.to_owned())
             .collect::<Vec<_>>()
             .join(", ")
     );
@@ -85,7 +85,7 @@ async fn init_bookmarks(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{MockCache, MockClient, Source};
+    use crate::{bookmarks::RawSource, MockCache, MockClient};
     use std::{
         collections::{HashMap, HashSet},
         path::Path,
@@ -96,7 +96,7 @@ mod tests {
         let client = MockClient::new();
         let cache = MockCache::new(CacheMode::Html);
         let bookmark_path = Path::new("test_data/bookmarks_chromium.json");
-        let source = Source::new(bookmark_path, vec![]);
+        let source = RawSource::new(bookmark_path, vec![]);
         let source_reader = SourceReader::init(&source).unwrap();
         let max_concurrent_requests = 100;
         let expected_bookmarks: HashSet<String> = HashSet::from_iter([
@@ -157,7 +157,7 @@ mod tests {
         let client = MockClient::new();
         let cache = MockCache::new(CacheMode::Text);
         let bookmark_path = Path::new("test_data/bookmarks_chromium.json");
-        let source = Source::new(bookmark_path, vec![]);
+        let source = RawSource::new(bookmark_path, vec![]);
         let source_reader = SourceReader::init(&source).unwrap();
         let max_concurrent_requests = 100;
         let expected_bookmarks: HashSet<String> = HashSet::from_iter([